+++ CHANGELOG.md
@@ -4,10 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.29.2]
 ## Added
 - Introduced multi-format writing for capture files. When format is 'multi' data
   will be written in both jsonl and parquet formats.
+- Added a 'Datadog Stateful Logs' blackhole for a new protocol we are testing
+  with the agent.
 
 ## [0.29.1]
 ## Fixed
@@ -32,7 +34,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a 'Datadog Intake' blackhole, supporting only the v2 metrics protobuf
   protocol for now. This captures metrics submitted into capture files.
 - Added the ability for lading to write capture files in parquet format.
-- Added a 'Datadog Stateful Logs' blackhole for a new protocol we are testing with the agent.
 
 ## [0.28.0]
 ## Added